cleanup and move status_update into base user class
