+++ src/elfpy/utils/price.py
@@ -95,8 +95,9 @@ def calc_liquidity(
 def calc_apr_from_spot_price(price, normalized_days_remaining):
     """Returns the APY (decimal) given the current (positive) base asset price and the remaining pool duration"""
     assert (
-        price > 0
-    ), f"pricing_models.calc_apy_from_spot_price: ERROR: calc_apy_from_spot_price: Price argument should be greater than zero, not {price}"
+        price >= 0
+    ), f"pricing_models.calc_apy_from_spot_price: ERROR: calc_apy_from_spot_price:"\
+        + f"Price argument should be greater or equal to zero, not {price}"
     assert (
         normalized_days_remaining > 0
     ), f"normalized_days_remaining argument should be greater than zero, not {normalized_days_remaining}"
